requirements.txt
Flask
flask-cors
openai
numpy
faiss-cpu
